src/python/requirements.txt
boto==2.38.0
boto3==1.2.4
Flask==0.10.1
flask_jsontools==0.1.1-0
Flask-SQLAlchemy==2.0
jsonpatch==1.11
jsonschema==2.5.1
mailer==0.8.1
objectpath==0.5
pinject==0.10.2
psycopg2==2.6
python-dateutil==2.4.2
pytz==2015.7
PyYAML==3.11
requests==2.7.0
retrying==1.3.3
uwsgi==2.0.10
sqlalchemy-redshift==0.4.0